Updated how banner info is stored and generated.
Moved the banner types struct to the BannersInfo file.

We will now keep track of JP and global dates in one array, instead of two. As this will make it easier to implement the overrides.

Updated the banner info calcs to work with this new set up.

Updated a bunch of variable names to reflect this change.