Item selection updates.
Made select2 selections wrap and will no longer display the dates.

Moved the select2 config into a variable.

When a banner is selected, the select2 will now close and remove all options from other banners. If no banner is selected than all options will be added back.

If an option is selected/deselected but the banner does not change, then the select2 will remain open and the options will not change.

Added a debounce for the resizing function.

Compile form will ignore rows with empty dropdowns.